+++ NEWS.md
@@ -3,6 +3,7 @@ Version 1.8.0.9000
 
 ### NEW FEATURES
 
+* [#1042](https://github.com/tidyverse/lubridate/issues/1042) `as_date` with character inputs accepts multiple formats in `format` argument. When `format` is supplied, the input string is parsed with `parse_date_time` instead of the old `strptime`.
 * [#1055](https://github.com/tidyverse/lubridate/issues/1055) Implement `as.integer` method for Duration, Period and Interval classes.
 * [#1061](https://github.com/tidyverse/lubridate/issues/1061) Make `year<-`, `month<-` etc. accessors truly generic. In order to make them work with arbitrary class XYZ, it's enough to define a `reclass_date.XYZ` method.
 * [#1061](https://github.com/tidyverse/lubridate/issues/1061) Add support for `year<-`, `month<-` etc. accessors for `data.table`'s IDate and ITime objects.